refactor: Cart 관련 로직 cart.js로 분리
